+++ lib/helpers/sample/sample-size.helper.ts
@@ -1,0 +1,6 @@
+/**
+ * $Size Payload
+ * @param size positive integer. Default to 1
+ * @constructor
+ */
+export const SampleSizeHelper = (size = 1) => ({ size });