+++ lib/interfaces/core/get-result.response.ts
@@ -1,13 +1,19 @@
 /**
  * The interface that represent the contract for a default GetResult response
  */
-export interface GetResultResponse {
+export interface GetResultResponse<T> {
   /**
    * The method used to obtain the documents found
-   * @param element Optional. Contains the index of the desired element or 'last' for the last element
    * @constructor
    */
-  GetDocs<T>(element?: number | 'last'): T;
+  GetDocs(): T[];
+
+  /**
+   * The method used to obtain only one element
+   * @constructor
+   * @param index
+   */
+  GetElement(index: number | 'last'): T;
 
   /**
    * The method used to obtain the total number of documents found