+++ lib/interfaces/index.ts
@@ -2,32 +2,32 @@
 /*
 CORE INTERFACES
  */
-export { BuilderOptionsInterface } from './core/builder-options.interface';
-export { InitOptionsInterface } from './core/init-options.interface';
-export { DebugBuildInterface } from './core/debug-build.interface';
-export { DebugHistoryInterface } from './core/debug-history.interface';
-export { StageInterface } from './core/stage.interface';
-export { StageErrorInterface } from './core/stage-error.interface';
-export { StageTypeEnum, getStageTypeValue, StageLabel } from './core/stage-type.interface';
+export { PipelineBuilderOptions } from './core/pipeline-builder-options';
+export { PipelineInitOptions } from './core/pipeline-init-options';
+export { DebugPipelineBuild } from './core/debug-pipeline-build';
+export { DebuggedAction } from './core/debugged-action';
+export { PipeLineStage } from './core/pipeline-stage';
+export { PipelineStageError } from './core/pipeline-stage-error';
+export { PipelineStageTypeEnum, getPipelineStageTypeValue, PipelineStageLabel } from './core/pipeline-stage-type';
 /*
 STAGE INTERFACES
  */
-export { BucketAutoStageInterface } from './stages/bucket-auto-stage.interface';
-export { BucketStageInterface } from "./stages/bucket-stage.interface";
-export { CollStatsStageInterface } from './stages/coll-stats-stage.interface';
-export { CurrentOpInterface } from './stages/current-op.interface';
-export { FacetStageInterface } from './stages/facet-stage.interface';
-export { GeoNearStageInterface } from './stages/geo-near-stage.interface';
-export { GraphLookupStageInterface } from './stages/graph-lookup-stage.interface';
-export { GroupStageInterface } from './stages/group-stage.interface';
-export { LookupConditionInterface } from './stages/lookup-condition.interface';
-export { LookupEqualityInterface } from './stages/lookup-equality.interface';
-export { LookupStageInterface } from './stages/lookup-stage.interface';
-export { MergeStageInterface, WhenMatchedType, WhenNotMatchedType } from './stages/merge-stage.interface';
-export { OutStageInterface } from './stages/out-stage.interface';
-export { ReplaceRootStageInterface } from './stages/replace-root-stage.interface';
-export { SampleStageInterface } from './stages/sample-stage.interface';
-export { UnionWithStageInterface } from './stages/union-with-stage.interface';
-export { UnwindStageInterface } from './stages/unwind-stage.interface';
+export { BucketAutoStage } from './stages/bucket-auto-stage';
+export { BucketStage } from "./stages/bucket-stage";
+export { CollStatsStage } from './stages/coll-stats-stage';
+export { CurrentOp } from './stages/current-op';
+export { FacetStage } from './stages/facet-stage';
+export { GeoNearStage } from './stages/geo-near-stage';
+export { GraphLookupStage } from './stages/graph-lookup-stage';
+export { GroupStage } from './stages/group-stage';
+export { LookupCondition } from './stages/lookup-condition';
+export { LookupEquality } from './stages/lookup-equality';
+export { LookupStage } from './stages/lookup-stage';
+export { MergeStage, WhenMatchedType, WhenNotMatchedType } from './stages/merge-stage';
+export { OutStage } from './stages/out-stage';
+export { ReplaceRootStage } from './stages/replace-root-stage';
+export { SampleStage } from './stages/sample-stage';
+export { UnionWithStage } from './stages/union-with-stage';
+export { UnwindStage } from './stages/unwind-stage';