+++ lib/methods/get-result/get-result.test.ts
@@ -1,6 +1,6 @@
-import { GetResult, getTotalPageNumber } from './get-result';
+import { GetPagingResult, GetResult, getTotalPageNumber } from './get-result';
 import { PipelineError } from '../../errors';
-import { StageInterface } from '../../interfaces';
+import { PipeLineStage } from '../../interfaces';
 import { GetResultResponse } from '../../interfaces/core/get-result.response';
 import { GetPagingResultResponse } from "../../interfaces/core/get-paging-result.response";
 
@@ -22,10 +22,13 @@ const pagingResultMock = [{
 }];
 
 const targetAggregationMock = {
-  aggregate: () => simpleResultMock,
+  aggregate: () => simpleResultMock.map((obj) => ({ ...obj })),
 };
 const targetPagingAggregationMock = {
-  aggregate: () => pagingResultMock,
+  aggregate: () => [{
+    docs: pagingResultMock[0].docs.map((obj) => ({ ...obj })),
+    count: pagingResultMock[0].count,
+  }],
 };
 
 const pagingPipelineMock = [{
@@ -38,7 +41,6 @@ const pagingPipelineMock = [{
 
 const pagingCountMock = [{ totalElements: 21 }];
 const pagingCountMock2 = [{ totalElements: 20 }];
-
 
 describe('GetResult', () => {
   it('should throw a PipelineError message if the target is not valid', async () => {
@@ -63,7 +65,7 @@ describe('GetResult', () => {
 
   it('should throw a PipelineError message if the pipeline is not valid', async () => {
     await expect(
-      () => GetResult(targetAggregationMock, undefined as unknown as StageInterface[])
+      () => GetResult(targetAggregationMock, undefined as unknown as PipeLineStage[])
     )
     .rejects
     .toThrowError(
@@ -83,7 +85,7 @@ describe('GetResult', () => {
 
   it('should throw a PipelineError message if one pipeline stage is unknown or not valid', async () => {
     await expect(
-      () => GetResult(targetAggregationMock, [{ $test: 'unit' } as StageInterface])
+      () => GetResult(targetAggregationMock, [{ $test: 'unit' } as PipeLineStage])
     )
       .rejects
       .toThrowError(
@@ -93,7 +95,7 @@ describe('GetResult', () => {
 
   it('should throw a PipelineError message if many pipeline stages are unknown or not valid', async () => {
     await expect(
-      () => GetResult(targetAggregationMock, [{ $test: 'unit' } as StageInterface, { $name: 'toto' } as StageInterface])
+      () => GetResult(targetAggregationMock, [{ $test: 'unit' } as PipeLineStage, { $name: 'toto' } as PipeLineStage])
     )
       .rejects
       .toThrowError(
@@ -118,38 +120,116 @@ describe('GetResult', () => {
   });
 
   it('should return an object with the GetDocs and GetCount methods as properties that get the documents (or the desired element if specified) and the total number of documents for a simple aggregate', async () => {
-    const getResult = await GetResult(targetAggregationMock, [{ $match: {} }]) as GetResultResponse;
+    const getResult = await GetResult(targetAggregationMock, [{ $match: {} }]) as GetResultResponse<any>;
     expect(getResult)
       .toHaveProperty('GetDocs');
     expect(getResult)
       .toHaveProperty('GetCount');
     expect(getResult.GetDocs())
       .toEqual(simpleResultMock);
-    expect(getResult.GetDocs(2))
+    expect(getResult.GetElement(2))
       .toEqual(simpleResultMock[2]);
-    expect(getResult.GetDocs(10000))
-      .toEqual(simpleResultMock[simpleResultMock.length - 1]);
-    expect(getResult.GetDocs('last'))
-      .toEqual(simpleResultMock[simpleResultMock.length - 1]);
+    expect(getResult.GetElement(10000))
+      .toEqual(undefined);
+    expect(getResult.GetElement('last'))
+      .toEqual(simpleResultMock.pop());
     expect(getResult.GetCount())
-      .toEqual(5);
+      .toEqual(4);
+  });
+});
+
+describe('GetPagingResult', () => {
+  it('should throw a PipelineError message if the target is not valid', async () => {
+    await expect(
+      () => GetPagingResult(undefined, [])
+    )
+    .rejects
+    .toThrowError(
+      new PipelineError('Application not possible, the target is not valid.')
+    );
+  });
+
+  it('should throw a PipelineError message if the target does not have the aggregate method', async () => {
+    await expect(
+      () => GetPagingResult({}, [])
+    )
+    .rejects
+    .toThrowError(
+      new PipelineError('Application not possible, the aggregate method does not exist on the chosen target.')
+    );
+  });
+
+  it('should throw a PipelineError message if the pipeline is not valid', async () => {
+    await expect(
+      () => GetPagingResult(targetAggregationMock, undefined as unknown as PipeLineStage[])
+    )
+    .rejects
+    .toThrowError(
+      new PipelineError('Application not possible, the pipeline is not valid.')
+    );
+  });
+
+  it('should throw a PipelineError message if the pipeline is empty', async () => {
+    await expect(
+      () => GetPagingResult(targetAggregationMock, [])
+    )
+    .rejects
+    .toThrowError(
+      new PipelineError('Application not possible, the pipeline is empty.')
+    );
+  });
+
+  it('should throw a PipelineError message if one pipeline stage is unknown or not valid', async () => {
+    await expect(
+      () => GetPagingResult(targetAggregationMock, [{ $test: 'unit' } as PipeLineStage])
+    )
+    .rejects
+    .toThrowError(
+      new PipelineError('Application not possible, $test pipeline stage is unknown or not valid.')
+    );
+  });
+
+  it('should throw a PipelineError message if many pipeline stages are unknown or not valid', async () => {
+    await expect(
+      () => GetPagingResult(targetAggregationMock, [{ $test: 'unit' } as PipeLineStage, { $name: 'toto' } as PipeLineStage])
+    )
+    .rejects
+    .toThrowError(
+      new PipelineError('Application not possible, $test / $name pipeline stages are unknown or not valid.')
+    );
+  });
+
+  it('should throw a PipelineError message if the MongoDb operation failed', async () => {
+    await expect(
+      () => GetPagingResult({
+        aggregate: async () => {
+          throw new Error('MongoDb aggregate method error.');
+        }
+      }, [{ $match: {} }])
+    )
+    .rejects
+    .toThrowError(
+      new PipelineError(
+        `An error was encountered while executing the GetPagingResult method:\n - MongoDb aggregate method error.`
+      )
+    );
   });
 
   it('should return an object with the GetDocs and GetCount methods as properties that get the documents and' +
     'the total number of documents for a paging aggregate', () => {
-    GetResult(targetPagingAggregationMock, [{ $match: {} }])
-      .then((getPagingResult) => {
-        expect(getPagingResult)
-          .toHaveProperty('GetDocs');
-        expect(getPagingResult)
-          .toHaveProperty('GetCount');
-        expect(getPagingResult.GetDocs())
-          .toEqual(pagingResultMock[0].docs);
-        expect(getPagingResult.GetCount())
-          .toEqual(pagingResultMock[0].count[0].totalElements);
-        expect((getPagingResult as GetPagingResultResponse).GetTotalPageNumber())
-        .toEqual(0);
-      });
+    GetPagingResult(targetPagingAggregationMock, [{ $match: {} }])
+    .then((getPagingResult) => {
+      expect(getPagingResult)
+      .toHaveProperty('GetDocs');
+      expect(getPagingResult)
+      .toHaveProperty('GetCount');
+      expect(getPagingResult.GetDocs())
+      .toEqual(pagingResultMock[0].docs);
+      expect(getPagingResult.GetCount())
+      .toEqual(pagingResultMock[0].count[0].totalElements);
+      expect((getPagingResult as GetPagingResultResponse<any>).GetTotalPageNumber())
+      .toEqual(-1);
+    });
   });
 });
 
@@ -169,7 +249,7 @@ describe('getTotalPageNumber', () => {
     pipeline
   ) => {
     expect(
-      getTotalPageNumber(count as unknown as any[], pipeline as unknown as StageInterface[])
+      getTotalPageNumber(count as unknown as any[], pipeline as unknown as PipeLineStage[])
     ).toEqual(-1);
   });