+++ lib/operators/boolean/index.test.ts
@@ -1,14 +1,13 @@
-import {And, Not, Or} from "./index";
+import {$And, $Not, $Or} from "./index";
 
-const
-    args: any[] = [],
-    expression = '';
+const args: any[] = [];
+const expression = '';
 
 describe('boolean operators', () => {
     test.each([
-        [And(...args), { $and: args }],
-        [Not(expression), { $not: [ expression ] }],
-        [Or(...args), { $or: args }],
+        [$And(...args), { $and: args }],
+        [$Not(expression), { $not: [expression ] }],
+        [$Or(...args), { $or: args }],
     ])('should return %s', (
         operation: any,
         expected: any