+++ lib/operators/comparison/index.ts
@@ -13,7 +13,7 @@
  * @param b
  * @constructor
  */
-export const Compare = (a: any, b: any) => ({ $cmp: [ a, b ] });
+export const $Compare = (a: any, b: any) => ({ $cmp: [a, b ] });
 /**
  * Compares two values and returns:
  * - true when the values are equivalent.
@@ -23,7 +23,7 @@ export const Compare = (a: any, b: any) => ({ $cmp: [ a, b ] });
  * @param b
  * @constructor
  */
-export const Equal = (a: any, b: any) => ({ $eq: [ a, b] });
+export const $Equal = (a: any, b: any) => ({ $eq: [a, b] });
 /**
  * Compares two values and returns:
  * - true when the first value is greater than the second value.
@@ -33,7 +33,7 @@ export const Equal = (a: any, b: any) => ({ $eq: [ a, b] });
  * @param b
  * @constructor
  */
-export const GreaterThan = (a: any, b: any) => ({ $gt: [ a, b ] });
+export const $GreaterThan = (a: any, b: any) => ({ $gt: [a, b ] });
 /**
  * Compares two values and returns:
  * - true when the first value is greater than or equivalent to the second value.
@@ -43,7 +43,7 @@ export const GreaterThan = (a: any, b: any) => ({ $gt: [ a, b ] });
  * @param b
  * @constructor
  */
-export const GreaterThanEqual = (a: any, b: any) => ({ $gte: [ a, b ] });
+export const $GreaterThanEqual = (a: any, b: any) => ({ $gte: [a, b ] });
 /**
  * Compares two values and returns:
  * - true when the first value is less than the second value.
@@ -53,7 +53,7 @@ export const GreaterThanEqual = (a: any, b: any) => ({ $gte: [ a, b ] });
  * @param b
  * @constructor
  */
-export const LessThan = (a: any, b: any) => ({ $lt: [ a, b ] });
+export const $LessThan = (a: any, b: any) => ({ $lt: [a, b ] });
 /**
  * Compares two values and returns:
  * - true when the first value is less than or equivalent to the second value.
@@ -63,7 +63,7 @@ export const LessThan = (a: any, b: any) => ({ $lt: [ a, b ] });
  * @param b
  * @constructor
  */
-export const LessThanEqual = (a: any, b: any) => ({ $lte: [ a, b] });
+export const $LessThanEqual = (a: any, b: any) => ({ $lte: [a, b] });
 /**
  * Compares two values and returns:
  * - true when the values are not equivalent.
@@ -73,4 +73,4 @@ export const LessThanEqual = (a: any, b: any) => ({ $lte: [ a, b] });
  * @param b
  * @constructor
  */
-export const NotEqual = (a: any, b: any) => ({ $ne: [ a, b] });
+export const $NotEqual = (a: any, b: any) => ({ $ne: [a, b] });