+++ lib/validators/stages/collStats-payload.validator.test.ts
@@ -1,5 +1,5 @@
 import {collStatsPayloadValidator} from "./collStats-payload.validator";
-import {CollStatsStageInterface} from "../../interfaces";
+import {CollStatsStage} from '../../interfaces';
 
 describe('collStats validators', () => {
     describe('collStatsPayloadValidator', () => {
@@ -11,13 +11,13 @@ describe('collStats validators', () => {
         ];
 
         test.each([
-            [collStatsPayloadValidator(payloadList[0] as unknown as CollStatsStageInterface),
+            [collStatsPayloadValidator(payloadList[0] as unknown as CollStatsStage),
             'The following mandatory fields are missing: latencyStats, storageStats, count, queryExecStats.'],
-            [collStatsPayloadValidator(payloadList[1] as unknown as CollStatsStageInterface),
+            [collStatsPayloadValidator(payloadList[1] as unknown as CollStatsStage),
                 'The histograms latencyStats property is missing.'],
-            [collStatsPayloadValidator(payloadList[2] as unknown as CollStatsStageInterface),
+            [collStatsPayloadValidator(payloadList[2] as unknown as CollStatsStage),
                 'The scale storageStats property is missing.'],
-            [collStatsPayloadValidator(payloadList[3] as unknown as CollStatsStageInterface),
+            [collStatsPayloadValidator(payloadList[3] as unknown as CollStatsStage),
                 ''],
         ])('%o should return %o', (
             operation: any,