+++ lib/validators/stages/graphLookup-payload.validator.ts
@@ -1,9 +1,9 @@
-import {GraphLookupStageInterface} from "../../interfaces";
+import {GraphLookupStage} from '../../interfaces';
 /**
  * Checks the presence of mandatory fields and the validity of each field present in the payload
  * @param payload The value passed to the stage
  * @returns an error message if non-compliant, an empty string if compliant
  */
-export const graphLookupPayloadValidator = (payload: GraphLookupStageInterface) => {
+export const graphLookupPayloadValidator = (payload: GraphLookupStage) => {
     return '';
 }