+++ lib/validators/stages/group-payload.validator.ts
@@ -1,10 +1,10 @@
-import {GroupStageInterface} from "../../interfaces";
+import {GroupStage} from '../../interfaces';
 
 /**
  * Checks the presence of mandatory fields and the validity of each field present in the payload
  * @param payload The value passed to the stage
  * @returns an error message if non-compliant, an empty string if compliant
  */
-export const groupPayloadValidator = (payload: GroupStageInterface) => {
+export const groupPayloadValidator = (payload: GroupStage) => {
     return '';
 }