+++ lib/validators/stages/out-payload.validator.ts
@@ -1,10 +1,10 @@
-import {OutStageInterface} from "../../interfaces";
+import {OutStage} from '../../interfaces';
 
 /**
  * Checks the presence of mandatory fields and the validity of each field present in the payload
  * @param payload The value passed to the stage
  * @returns an error message if non-compliant, an empty string if compliant
  */
-export const outPayloadValidator = (payload: OutStageInterface) => {
+export const outPayloadValidator = (payload: OutStage) => {
     return '';
 }